[CALCITE-2569] UDFs that are table functions must implement SqlTableFunction and have CURSOR as their return type
Validate that table functions are not used in contexts that require
scalar functions, such as the SELECT clause.

Change operand type ANY to mean 'scalar expression of any type (but
not a cursor)', and add operand type IGNORE to mean skip validation -
for an operand that is not an expression. TABLE is one of the few
operators that accepts a CURSOR operand.