Prototype [CALCITE-3923] Refactor how planner rules are parameterized
I've only tackled two rules: ProjectFilterTransposeRule and
its sub-class DruidFilterTransposeRule.

The prototype demonstrates that in order to modify a rule,
you can take the config of an existing rule instance, alter
some properties, optionally convert it to a sub-class config,
then convert the config to a rule.

I had to add a little magic to ImmutableBeans, but not much.

The change is backwards compatible, in the sense that
existing rule constructors are deprecated but still work. The
way to create rules in future will be via
RelOptRule.Config.toRule() (which calls the rule's (Config)
constructor).