[CALCITE-4366] LatticeSuggester should treat UNION and VALUES sub-expressions as tables
We achieve this by adding a wrapper table function in the
RelNode tree above any UNION or VALUES after sql-to-rel but
before calling LatticeSuggester.

If the UNION or VALUES expression is a common-table expression
(i.e. in the WITH clause), use the name of the WITH clause
table and columns for the dummy table. That allows use to
generate a nicer model later.

Add SqlToRelConverter.Config.postStep(), to apply a
user-specified transform to a RelNode after sql-to-rel
conversion. (We could make decorrelation use this hook.)

Implement getCursorOperand(int) in RexCallBinding, so that
when you call RelBuilder.tableFunctionScan the row-type can
be deduced in terms of the relational input(s).

Refactor LatticeSuggesterTest to make it easier to write
tests with custom config.