fixup
In CalciteResource, rename offsetStartLimitCountNotAllowed to offsetLimitNotAllowed

In SqlConformance, rename isOffsetStartLimitCountAllowed to isOffsetLimitAllowed

In reference, simplify the description of the order of LIMIT, OFFSET, FETCH clause

In parser, rename startCount to offsetFetch (based on the standard names of the clauses). Extract FETCH into its own method. Disallow FETCH if parser has already seen LIMIT.