almost there
The following functions allow time frame arguments:
 * DATEADD (Postgres, MSSql)
 * DATEDIFF (Postgres, MSSql)
 * DATEPART (MSSql)
 * DATE_PART (Postgres)
 * EXTRACT (Calcite built-in, also SQL standard)
 * CEIL (Calcite built-in)
 * FLOOR (Calcite built-in)
 * TIMESTAMPADD (Calcite built-in, also JDBC standard)
 * TIMESTAMPDIFF (Calcite-builtin, also JDBC standard)

Calls to the above functions with invalid time units would
previously be a parse error and are now detected during
validation.

The SQL_TSI_xxx (e.g. SQL_TSI_HOUR) arguments are treated
as time frames, and the parser passes them as identifiers.
They are no longer reserved keywords.

Previously, NANOSECOND and MILLISECOND were allowed in
EXTRACT but no other functions.

The representation of calls to the above functions has
changed. Previously the operand was a time unit, now it is
an identifier.

Add a new function table "MSSQL" (for Microsoft SQL Server).