Refactor RexImpTable
Break up the constructor into a builder pattern, and make
RexImpTable immutable after construction.

The constructor was getting too long (approaching 400
lines), so this change introduces a private static class
Builder and moves initialization into chained populate()
and populate2() methods.

Usually we put inner classes towards the end of the file.
In this case, we put the Builder inner class after the
constructor, in order to reduce diff noise.