[CALCITE-5557] Add SAFE_CAST function (enabled in BigQuery library)
SAFE_CAST is similar to CAST, except that it returns NULL
rather than throwing an error if conversion fails.

The Java code generation generates the same code as for CAST,
wraps it as a lambda, and converts the exception to Java null.

Add a wrapper around SqlOperatorFixture so that existing
tests for CAST can also test SAFE_CAST (Julian Hyde).

Close apache/calcite#3093

Co-authored-by: Oliver Lee <oliverlee@google.com>
Co-authored-by: Julian Hyde <jhyde@apache.org>