RelBuilder.aggregate now adds an aggregate function to ensure that there is at least one field
Solves that problem that an Aggregate query that uses no
columns throws AssertionError in RelFieldTrimmer.trimFields.

Add an option RelBuilder.Config.preventEmptyFieldList to
turn off the behavior we just added.

If, in future, we implement "[CALCITE-4334] TRUE_AGG, an
aggregate function that always returns TRUE", we should use
that rather than "COUNT(*)".