[CALCITE-5764] Add Puffin
Puffin provides a similar programming model to Awk. A program
is a collection of rules, each of which is a predicate
applied to a line of text followed by an action on that test.

Puffin programs are thread-safe; if Puffin is invoked on
mutiple sources (files), each file is processed in its own
thread and is allocated its own state that the rules can use
without coordination.

Puffin caches compiled regular expressions, so that calls to
`matches(String)` are almost as efficient as if the program
had called `Pattern.compile(String)` in advance.