Refactor parser
In SqlAbstractParserImpl, make normalizeException method non-abstract,
and add methods normalizeParseException and normalizeTokenMgrError.
These methods contain logic that was previously in the generated parser.

In MetadataImpl, use immutable lists and sets.

Cache metadata in the factory, rather than in a static field in the
generated parser class. To this end, add method
SqlParserImplFactory.getMetadata().