Refactor parser
In SqlAbstractParserImpl, make normalizeException method non-abstract,
and add methods normalizeParseException and normalizeTokenMgrError.
These methods contain logic that was previously in the generated parser.

In `MetadataImpl`, use immutable lists and sets.

Cache metadata in the factory, rather than in a static field in the
generated parser class. To this end, add method
`SqlParserImplFactory.getMetadata()`.

The Babel parser needs more lookahead, but setting JavaCC's lookahead
globally to 2 causes it to generate syntactically incorrect Java
(see javacc/javacc#288). So, introduce the `${parser.babelLa2}` FMPP
variable and use it in `Parser.jj` wherever the Babel parser needs
`LOOKAHEAD(2)` and the regular parser does not.