Revise
In SET clause of UPDATE statement, allow column identifiers to be prefixed with table alias

For example, in the following query 'e' is a table alias and
'e.deptno' is the column name prefixed with the table alias.
Previously only 'deptno' was allowed.

  UPDATE emp AS e
    SET e.deptno = 20;

A follow-up case, [CALCITE-6584], will implement this in the
validator.

Close apache/calcite#3959