[CALCITE-6598] In RelDataTypeSystem, deprecate methods getMaxNumericScale and getMaxNumericPrecision
In release 1.38, we leave in place the implementation of these
methods in RelDataTypeSystemImpl and DelegatingTypeSystem, in case
a project has overridden these methods.

If a project has overridden either of these methods, it should
remove the overrides and instead override getMaxScale or
getMaxPrecision before upgrading to release 1.39.

In release 1.39, we shall make these methods final, which will break
any project that has overridden them.